Split models and views
